move training control buttons to their own file
