+++ packages/client/src/components/project/input/Input.tsx
@@ -1,14 +1,9 @@
-import { ResumeTraining, ResumeTrainingVariables } from "./__generated__/ResumeTraining";
-import { StartTraining, StartTrainingVariables } from "./__generated__/StartTraining";
-import { PauseTraining, PauseTrainingVariables } from "./__generated__/PauseTraining";
-import { StopTraining, StopTrainingVariables } from "./__generated__/StopTraining";
-import { GetDockerState } from "../../trainerStatus/__generated__/GetDockerState";
-import { DockerState, TrainStatus } from "../../../__generated__/globalTypes";
-import { GET_DOCKER_STATE } from "../../trainerStatus/TrainerStatus";
-import { Button, Container, Divider, LinearProgress, Typography, Box, LinearProgressProps } from "@material-ui/core";
+import { Container, Divider, LinearProgress, Typography, Box, LinearProgressProps } from "@material-ui/core";
+import { StartButton, StopButton, PauseButton } from "./Buttons";
+import { TrainStatus } from "../../../__generated__/globalTypes";
 import { GetTrainjobs } from "./__generated__/GetTrainjobs";
-import { gql, useMutation, useQuery } from "@apollo/client";
-import React, { ReactElement, useState } from "react";
+import { gql, useQuery } from "@apollo/client";
+import React, { ReactElement } from "react";
 import Parameters from "./Parameters";
 
 const GET_TRAINJOBS = gql`
@@ -22,43 +17,7 @@ const GET_TRAINJOBS = gql`
   }
 `;
 
-const START_TRAINING = gql`
-  mutation StartTraining($id: ID!) {
-    startTraining(id: $id) {
-      id
-    }
-  }
-`;
-
-const STOP_TRAINING = gql`
-  mutation StopTraining($id: ID!) {
-    stopTraining(id: $id) {
-      id
-    }
-  }
-`;
-
-const PAUSE_TRAINING = gql`
-  mutation PauseTraining($id: ID!) {
-    pauseTraining(id: $id) {
-      id
-    }
-  }
-`;
-
-const RESUME_TRAINING = gql`
-  mutation ResumeTraining($id: ID!) {
-    resumeTraining(id: $id) {
-      id
-    }
-  }
-`;
-
 export default function Input(props: { id: string }): ReactElement {
-  const id = props.id;
-  //^^ this is only to get around a props validation error when using
-  //   the props in the function components defined within this function component. need help with this.
-
   const { data, loading, error } = useQuery<GetTrainjobs>(GET_TRAINJOBS, {
     pollInterval: 2000
   });
@@ -72,11 +31,9 @@ export default function Input(props: { id: string }): ReactElement {
   if (trainjob === undefined)
     return (
       <>
-        {/*<Datasets id={props.id} />*/}
-        {/*<Divider />*/}
         <Parameters id={props.id} />
         <Divider />
-        <StartButton />
+        <StartButton id={props.id} />
       </>
     );
 
@@ -112,73 +69,11 @@ export default function Input(props: { id: string }): ReactElement {
     <Container>
       <h1>{statusMessage}</h1>
       <ProgressBar />
-      <StopButton />
-      <PauseButton />
+      <StopButton id={props.id} />
+      <PauseButton id={props.id} job={trainjob} />
     </Container>
   );
 
-  function StartButton(): ReactElement {
-    const [startTraining] = useMutation<StartTraining, StartTrainingVariables>(START_TRAINING);
-    const [starting, setStarting] = useState(false);
-
-    const handleClick = () => {
-      startTraining({ variables: { id: id } });
-      setStarting(true);
-    };
-
-    const { data, loading, error } = useQuery<GetDockerState>(GET_DOCKER_STATE, { pollInterval: 5000 });
-    if (loading) return <p>connecting to trainer</p>;
-    if (error) return <p>cant connect to trainer</p>;
-    if (data?.dockerState === DockerState.TRAIN_PULL) return <p>no train image yet</p>;
-    if (starting) return <Button>Starting...</Button>;
-
-    return (
-      <Button onClick={handleClick} color="primary" variant="contained">
-        Start
-      </Button>
-    );
-  }
-
-  function StopButton(): ReactElement {
-    const [stopTraining] = useMutation<StopTraining, StopTrainingVariables>(STOP_TRAINING);
-    const [stopping, setStopping] = useState(false);
-
-    const handleClick = () => {
-      stopTraining({ variables: { id: id } });
-      setStopping(true);
-    };
-    if (stopping) return <Button>Stopping...</Button>;
-
-    return <Button onClick={handleClick}>Stop</Button>;
-  }
-
-  function PauseButton(): ReactElement {
-    const [pauseTraining] = useMutation<PauseTraining, PauseTrainingVariables>(PAUSE_TRAINING);
-    const [resumeTraining] = useMutation<ResumeTraining, ResumeTrainingVariables>(RESUME_TRAINING);
-
-    const [pausing, setPausing] = useState(false);
-    const [resuming, setResuming] = useState(false);
-
-    const handlePause = () => {
-      pauseTraining({ variables: { id: id } });
-      setPausing(true);
-    };
-
-    const handleResume = () => {
-      resumeTraining({ variables: { id: id } });
-      setResuming(true);
-    };
-
-    if (trainjob?.status === TrainStatus.Paused) {
-      if (pausing) setPausing(false);
-      if (resuming) return <Button>Resuming...</Button>;
-      return <Button onClick={handleResume}>Resume</Button>;
-    } else {
-      if (resuming) setResuming(false);
-      if (pausing) return <Button>Pausing...</Button>;
-      return <Button onClick={handlePause}>Pause</Button>;
-    }
-  }
   function LinearProgressWithLabel(props: LinearProgressProps & { value: number }) {
     return (
       <Box display="flex" alignItems="center">
